Add local storage inspector integration Moved all inspectors to single directory Upgraded all dependencies Update to flutter 2.8.1
