+++ .fvm/fvm_config.json
@@ -1,4 +1,4 @@
 {
-  "flutterSdkVersion": "2.8.0",
+  "flutterSdkVersion": "2.8.1",
   "flavors": {}
 }
