+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
   scroll_when_needed: ^2.0.0
   shared_preferences: ^2.0.11
   sqlite3_flutter_libs: ^0.5.2
+  storage_inspector: ^0.1.1
 
 dev_dependencies:
   build_runner: ^2.1.1
@@ -46,7 +47,7 @@ dev_dependencies:
   json_serializable: ^6.1.1
   mockito: ^5.0.15
   model_generator: ^5.8.1
-  retrofit_generator: 3.0.0+2
+  retrofit_generator: ^3.0.1+1
 
 flutter:
   uses-material-design: true
