+++ test/database/flutter_database_inspector_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_template/util/database/database_inspector.dart';
+import 'package:flutter_template/util/inspector/database_inspector.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../di/test_injectable.dart';
